+++ Unity-Technologies-networking/Runtime/Transport/Transport.cs
@@ -7,7 +7,7 @@ namespace UnityEngine.Networking
     public static class Transport
     {
         // hlapi needs to know max packet size to show warnings
-        public static int MaxPacketSize = ushort.MaxValue;
+        public static int MaxPacketSize = int.MaxValue;
 
         // selected transport layer: Telepathy by default
         public static TransportLayer layer = new TelepathyWebsocketsMultiplexTransport();
